# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
  /model-catalog:
    get:
      tags: [models]
      summary: 获取创作端模型目录
      description: |
        固定 source=system，channels 直接来自系统渠道下的模型列表，models 为空数组；不反查逻辑模型。
        一级跨渠道按 displayName 分组，二级显示 channelLabel；模型请求仍使用 channelId + modelKey 定位。
        channelLabel 是后台渠道模型创建和更新时配置的渠道展示名，去除首尾空白后最多 80 字，空字符串表示使用所属渠道名称。
        description 是后台渠道模型创建和更新时填写的公开模型描述，去除首尾空白后最多 500 字；空字符串表示未填写。
        displayName 和 icon 用于一级模型菜单，channelLabel 不参与路由、权限或计费。
        渠道级 publicAlias 已废弃；成本配置和成本快照仅供管理端，不包含在本接口及用户报价中。
      responses:
        "200":
          description: 当前启用的模型目录
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    enum: [0]
                  msg:
                    type: string
                  data:
                    type: object
                    required: [source, models, channels]
                    properties:
                      source:
                        type: string
                        enum: [system]
                      models:
                        type: array
                        items:
                          type: object
                      channels:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            displayName:
                              type: string
                            models:
                              type: array
                              items:
                                type: object
                                properties:
                                  modelKey:
                                    type: string
                                  displayName:
                                    type: string
                                  icon:
                                    type: string
                                  channelLabel:
                                    type: string
                                    maxLength: 80
                                    default: ""
                                  description:
                                    type: string
                                    maxLength: 500
                                    default: ""
        "401":
          $ref: "#/components/responses/Error"
  /model-catalog/quote:
    post:
      tags: [models]
      summary: 按当前参数预估系统渠道模型费用
      description: |
        需要登录会话 Cookie。仅接受启用的系统渠道和模型，采用服务端能力默认值及价格档选择规则。
        不创建任务或账单、不冻结或扣除积分，也不调用模型供应商。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId, modelKey, intent]
              properties:
                channelId:
                  type: string
                  minLength: 1
                modelKey:
                  type: string
                  minLength: 1
                  description: 系统渠道中配置的模型键
                intent:
                  $ref: "#/components/schemas/ModelRequestIntent"
      responses:
        "200":
          $ref: "#/components/responses/ModelQuote"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
...
  /public/welcome:
    get:
      tags: [meta]
      security: []
      summary: 读取欢迎页开放状态
      responses:
        "200":
          description: 信封 data 仅包含 welcomeEnabled 布尔值，默认 true；Cache-Control 为 no-store
        "500":
          description: 配置读取失败
...
  /agent/memories:
    get:
      tags: [agent]
      summary: 列出当前用户的 Agent 记忆
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: 信封 data.memories
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [agent]
      summary: 手动添加一条个人记忆（直接批准）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, category, situation]
              properties:
                topic: { type: string }
                category: { type: string }
                situation: { type: string }
                lesson: { type: string }
                source: { type: string }
                steps:
                  type: array
                  items:
                    type: object
                    properties:
                      tool: { type: string }
                      action: { type: string }
                      note: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/export:
    get:
      tags: [agent]
      summary: 导出当前用户的 Agent 记忆
      responses:
        "200":
          description: 信封 data 为 version/kind/exportedAt/memories
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/import:
    post:
      tags: [agent]
      summary: 导入 Agent 记忆 JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version: { type: integer }
                kind: { type: string }
                memories: { type: array }
      responses:
        "200":
          description: 信封 data 含 imported、merged、skipped
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/settings:
    get:
      tags: [agent]
      summary: 读取当前用户的记忆压缩设置与最近一次压缩状态
      responses:
        "200":
          description: 信封 data 含 compactInterval、模型字段、lastStatus、summary
        "401":
          $ref: "#/components/responses/Error"
    patch:
      tags: [agent]
      summary: 更新记忆压缩周期和文本模型
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compactInterval:
                  type: string
                  enum: [off, daily, weekly, monthly]
                logicalModelId: { type: string }
                channelId: { type: string }
                channelModelKey: { type: string }
                model: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/compact:
    post:
      tags: [agent]
      summary: 提交一次个人记忆压缩任务（走用户文本模型计费）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logicalModelId: { type: string }
                channelId: { type: string }
                channelModelKey: { type: string }
                model: { type: string }
      responses:
        "200":
          description: 信封 data 为压缩设置视图，lastStatus 为 queued
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /agent/memories/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [agent]
      summary: 更新一条属于当前用户的记忆
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [agent]
      summary: 删除一条属于当前用户的记忆
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /agent/memories/{id}/decide:
    post:
      tags: [agent]
      summary: 批准或拒绝一条属于当前用户的待审记忆
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [approve, reject]
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /admin/agent-lessons:
    get:
      tags: [admin]
      summary: 按用户筛选查看个人 Agent 记忆
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: 信封 data.lessons 含解析后的 steps、authorUserId、authorUsername、authorDisplayName
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /admin/agent-lessons/{id}:
    delete:
      tags: [admin]
      summary: 删除一条用户的 Agent 记忆
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
    ModelRequestIntent:
      type: object
      required: [capability]
      properties:
        capability:
          type: string
          enum: [text, image, video, audio]
        operation:
          type: string
          description: 可选生成操作，必须符合所选模型的能力合同
        inputs:
          type: object
          description: 按素材类型统计的参考数量，如 image、video、audio；另受所选模型的数量上限约束
          additionalProperties:
            type: integer
            minimum: 0
            maximum: 128
        options:
          type: object
          description: 当前创作规格；未指定的参数使用服务端默认值
          additionalProperties: true
          example:
            vquality: 720p
            size: "16:9"
            videoSeconds: 5
            videoGenerateAudio: true
    LogicalModelQuote:
      type: object
      required: [logicalModelId, billingMode, quantity, amountMicrocredits, estimated]
      properties:
        logicalModelId:
          type: string
          description: 前台逻辑模型 ID；直接查询系统渠道模型时为空字符串
        billingMode:
          type: string
          enum: [fixed_request, per_second, token]
        quantity:
          type: integer
          format: int64
          minimum: 1
          description: 按次为 1，按秒为生成秒数，按 Token 为预留用量（视频含平台预留余量）
        amountMicrocredits:
          type: integer
          format: int64
          minimum: 0
          description: 预估预留微积分，1000000 微积分等于 1 积分；明确配置的免费价格返回 0
        estimated:
          type: boolean
          description: Token 计费为 true；优先有效上游 usage，视频无有效用量时按提交的公式快照结算，预留不构成费用上限
        videoTokenEstimate:
          $ref: "#/components/schemas/VideoTokenEstimate"
    VideoTokenEstimate:
      type: object
      description: 仅视频 Token 报价返回；将火山公式用量和平台预留余量分开披露
      required: [formulaTokens, reservedTokens, outputWidth, outputHeight, framesPerSecond, outputSeconds, referenceSeconds, referenceDurationEstimated, dimensionsEstimated, reservationMarginPercent]
      properties:
        formulaTokens:
          type: integer
          format: int64
          minimum: 1
          description: 按（参考视频总秒数 + 输出秒数）× 输出宽 × 输出高 × 帧率 / 1024 向上取整；上游最低计费用量可能更高
        reservedTokens:
          type: integer
          format: int64
          minimum: 1
          description: 公式用量叠加平台预留余量后向上取整
        outputWidth:
          type: integer
          format: int64
          minimum: 1
        outputHeight:
          type: integer
          format: int64
          minimum: 1
        framesPerSecond:
          type: integer
          format: int64
          minimum: 1
        outputSeconds:
          type: integer
          format: int64
          minimum: 1
        referenceSeconds:
          type: number
          format: double
          minimum: 0
          description: 计入公式的参考视频总秒数；报价仅有参考数量时采用预估时长
        referenceDurationEstimated:
          type: boolean
          description: 参考时长未知、使用预估值时为 true
        dimensionsEstimated:
          type: boolean
          description: 自适应画幅或上游型号无法确定，使用预估像素尺寸时为 true
        reservationMarginPercent:
          type: integer
          format: int64
          minimum: 0
          description: 平台预留余量百分比，不属于火山官方计费公式
    ModelQuoteEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            code:
              type: integer
              enum: [0]
            data:
              type: object
              required: [quote]
              properties:
                quote:
                  $ref: "#/components/schemas/LogicalModelQuote"
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
    ModelQuote:
      description: 当前参数对应的费用报价，位于信封 data.quote
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ModelQuoteEnvelope"
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"